% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_eur_to_usd.R
\name{convert_usd_to_eur}
\alias{convert_usd_to_eur}
\title{Convert USD to EUR}
\usage{
convert_usd_to_eur(usd)
}
\arguments{
\item{usd}{number}
}
\value{
number
}
\description{
The function looks up the most recent USD/EUR exchange rate and computes the provided
USD amount in EUR.
}
\examples{
convert_usd_to_eur(100)
}
