% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusd.R
\name{eurusd}
\alias{eurusd}
\title{Get latest EUR-USD exchange rate}
\usage{
eurusd()
}
\value{
number
}
\description{
The function returns the latest Euro - US Dollar exchange rate, i.e. the latest price
of 1 USD in EUR.
}
