% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusds.R
\name{eurusds}
\alias{eurusds}
\title{Get EUR-USD exchange rates}
\usage{
eurusds(date_from, date_to)
}
\arguments{
\item{date_from}{date. Formatted as yyyy-mm-dd}

\item{date_to}{date. Formatted as yyyy-mm-dd}
}
\value{
data.table
}
\description{
The function gets EUR-USD exchange rates for the provided time period.
}
\examples{
eurusds('2019-05-20', '2019-05-27')
}
