% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uneur.R
\name{uneur}
\alias{uneur}
\title{Formats string in a Euro format to a number}
\usage{
uneur(x)
}
\arguments{
\item{x}{string. Should be in a Euro format}
}
\value{
number
}
\description{
The function is the reverse of \code{eur}.
}
\examples{
uneur("€1,000,000")
uneur("€-1.25")
}
\seealso{
eur
}
